requirements.txt
Flask==1.0.2
Flask-Admin==1.5.1
Flask-Bcrypt==0.7.1
Flask-Cors
Flask-Loginmanager==1.1.6
Flask-Login==0.4.1
Flask-mail==0.9.1
Flask-Migrate==2.2.0
Flask-Restplus==0.12.1
Flask-Script==2.0.6
Flask-Security==3.0.0
Flask-Session==0.3.1
Flask-Social==1.6.2
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
itsdangerous==0.24
jinja2==2.10
mysql-connector==2.1.6
urllib3==1.24.1
requests==2.21.0
bcrypt==3.1.4
pyOpenSSL==18.0
Flask-OAuthlib==0.9.5
python-jose==3.0.0
xmltodict==0.11.0